openff-toolkit/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
#amberlite=16.0=pypi_0
#ambertools=20.15=pypi_0
arpack=3.7.0=hdefa2d7_2
astunparse=1.6.3=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blosc=1.21.0=h9c3ff4c_0
boost=1.74.0=py37h6dcda5c_3
boost-cpp=1.74.0=hc6e9bd1_2
bzip2=1.0.8=h7f98852_4
c-ares=1.17.1=h7f98852_1
ca-certificates=2020.12.5=ha878542_0
cairo=1.16.0=h6cf1ce9_1008
certifi=2020.12.5=py37h89c1867_1
cudatoolkit=11.2.2=he111cf0_8
curl=7.76.0=h979ede3_0
cycler=0.10.0=py_2
cython=0.29.22=py37hcd2ae1e_0
decorator=5.0.7=pyhd8ed1ab_0
fftw=3.3.9=nompi_h74d3f13_101
fontconfig=2.13.1=hba837de_1005
freetype=2.10.4=h0708190_1
gettext=0.19.8.1=h0b5b191_1005
greenlet=1.0.0=py37hcd2ae1e_0
hdf4=4.2.13=h10796ff_1004
hdf5=1.10.6=nompi_h6a2412b_1114
icu=68.1=h58526e2_0
importlib-metadata=3.10.1=py37h89c1867_0
ipython=7.22.0=py37h085eea5_0
ipython_genutils=0.2.0=py_1
jedi=0.18.0=py37h89c1867_2
jpeg=9d=h36c2ea0_0
kiwisolver=1.3.1=py37h2527ec5_1
krb5=1.17.2=h926e7f8_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.35.1=hea4e1c9_2
libblas=3.9.0=8_openblas
libcblas=3.9.0=8_openblas
libcurl=7.76.0=hc4aaa36_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libffi=3.3=h58526e2_2
libgcc-ng=9.3.0=h2828fa1_19
libgfortran-ng=9.3.0=hff62375_19
libgfortran5=9.3.0=hff62375_19
libglib=2.68.1=h3e27bee_0
libgomp=9.3.0=h2828fa1_19
libiconv=1.16=h516909a_0
liblapack=3.9.0=8_openblas
libnetcdf=4.7.4=nompi_h56d31a8_107
libnghttp2=1.43.0=h812cca2_0
libopenblas=0.3.12=pthreads_h4812303_1
libpng=1.6.37=h21135ba_2
libssh2=1.9.0=ha56f1ee_6
libstdcxx-ng=9.3.0=h6de172a_19
libtiff=4.2.0=hdc55705_0
libuuid=2.32.1=h7f98852_1000
libwebp-base=1.2.0=h7f98852_2
libxcb=1.13=h7f98852_1003
libxml2=2.9.10=h72842e0_4
libxslt=1.1.33=h15afd5d_2
lxml=4.6.3=py37h77fd288_0
lz4-c=1.9.3=h9c3ff4c_0
lzo=2.10=h516909a_1000
matplotlib-base=3.4.1=py37hdd32ed1_0
mdtraj=1.9.5=py37hd0d7e5a_1
#mmpbsa-py=16.0=pypi_0
mock=4.0.3=py37h89c1867_1
ncurses=6.2=h58526e2_4
netcdf-fortran=4.5.3=nompi_h996563d_103
networkx=2.3=py_0
numexpr=2.7.3=py37hdc94413_0
numpy=1.20.2=py37h038b26d_0
ocl-icd=2.3.0=h7f98852_0
ocl-icd-system=1.0.0=1
olefile=0.46=pyh9f0ad1d_1
openbabel=3.1.1=py37h200e996_1
openff-forcefields=1.3.0=pyh44b312d_0
openff-toolkit=0.9.1=pyhd8ed1ab_2
openff-toolkit-base=0.9.1=pyhd8ed1ab_2
openjpeg=2.4.0=hf7af979_0
openmm=7.5.0=py37h1b3cdbe_6
openmmforcefields=0.9.0=pyhd8ed1ab_0
openssl=1.1.1k=h7f98852_0
packaging=20.9=pyh44b312d_0
#packmol-memgen=1.1.0rc0=pypi_0
pandas=1.2.4=py37h219a48f_0
parmed=3.4.1=py37hcd2ae1e_0
parso=0.8.2=pyhd8ed1ab_0
pcre=8.44=he1b5a44_0
#pdb4amber=1.7.dev0=pypi_0
perl=5.32.0=h36c2ea0_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.1.2=py37h4600e1f_1
pip=21.0.1=pyhd8ed1ab_0
pixman=0.40.0=h36c2ea0_0
prompt-toolkit=3.0.18=pyha770c72_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pycairo=1.20.0=py37h01af8b0_1
pygments=2.8.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pytables=3.6.1=py37h0c4f3e0_3
python=3.7.10=hffdb5ce_100_cpython
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
#pytraj=2.0.5=pypi_0
pytz=2021.1=pyhd8ed1ab_0
pyyaml=5.4.1=py37h5e8e339_0
rdkit=2021.03.1=py37haf5a968_0
readline=8.0=he28a2e2_2
reportlab=3.5.67=py37h69800bb_0
#sander=16.0=pypi_0
scipy=1.6.2=py37h29e03ee_0
setuptools=49.6.0=py37h89c1867_3
six=1.15.0=pyh9f0ad1d_0
smirnoff99frosst=1.1.0=pyh44b312d_0
snappy=1.1.8=he1b5a44_3
sqlalchemy=1.4.7=py37h5e8e339_0
sqlite=3.35.4=h74cdb3f_0
tinydb=4.4.0=pyh44b312d_0
tk=8.6.10=h21135ba_1
tornado=6.1=py37h5e8e339_1
traitlets=5.0.5=py_0
typing_extensions=3.7.4.3=py_0
wcwidth=0.2.5=pyh9f0ad1d_2
wheel=0.36.2=pyhd3deb0d_0
xmltodict=0.12.0=py_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.0=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-libxt=1.2.1=h7f98852_2
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h516909a_1
yaml=0.2.5=h516909a_0
zipp=3.4.1=pyhd8ed1ab_0
zlib=1.2.11=h516909a_1010
zstd=1.4.9=ha95c52a_0
